+++ src/main/java/com/capgemini/wsb/fitnesstracker/user/internal/UserController.java
@@ -1,14 +1,9 @@
 package com.capgemini.wsb.fitnesstracker.user.internal;
 
-import com.capgemini.wsb.fitnesstracker.training.api.Training;
 import com.capgemini.wsb.fitnesstracker.user.api.User;
 import com.capgemini.wsb.fitnesstracker.user.api.UserNotFoundException;
-import com.capgemini.wsb.fitnesstracker.user.api.UserService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.java.Log;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
@@ -48,7 +43,11 @@ class UserController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteUser(@PathVariable Long id) {
         Optional<User> userOptional = userService.getUserById(id);
-        userService.deleteUser(id);
+        if (userOptional.isPresent()) {
+            userService.deleteUser(id);
+        } else {
+            throw new UserNotFoundException(id);
+        }
     }
 
     @GetMapping("/{id}")
